--- src/components/getTripAdvisor/network.js
@@ -1,8 +1,0 @@
-import { Router } from 'express'
-import { getTrip } from './controller.js'
-
-const getTripRoute = Router()
-
-getTripRoute.get('/', getTrip)
-
-export { getTripRoute }
