--- src/components/lenguage/network.js
@@ -1,8 +1,0 @@
-import { Router } from 'express'
-import { getActivity } from './controller.js'
-
-const lenguageRouter = Router()
-
-lenguageRouter.post('/getActivity', getActivity)
-
-export { lenguageRouter }
