+++ src/utils/deberta/consumDeberta.js
@@ -1,4 +1,3 @@
-import { response } from '../../network/response.js'
 import city from './json/context-city.json' assert { type: 'json' }
 import key from './json/key.json' assert { type: 'json' }
 import actions from './json/action.json' assert { type: 'json' }
@@ -28,29 +27,25 @@ const listActivity = async (data) => {
   return placesAndActions
 }
 
-export const getActivity = async (req, res) => {
+export const getActivity = async (cityName, contextUser) => {
   try {
-    const { cityName, context_user } = req.body
-    // Use the search function to find the city
     const info = city.city
     const cityInfo = searchCity(info, cityName)
 
     if (!cityInfo) {
-      response.error(res, 'Ciudad no encontrada', 404)
+      return {
+        error: 'No se encontró la ciudad'
+      }
     } else {
-      if (context_user === undefined) {
+      if (contextUser === undefined) {
         response.error(res, 'Falta el contexto del usuario', 400)
       }
-      const activities = await lenguageProcess(cityInfo, context_user)
+      const activities = await lenguageProcess(cityInfo, contextUser)
       const toTodo = await listActivity(activities)
-      response.success(
-        res,
-        'Actividades encontradas',
-        { activities: toTodo },
-        200
-      )
+      return toTodo
     }
   } catch (error) {
-    response.error(res, error.message, 500)
+    console.error('Ha ocurrido un error:', error)
+    throw error
   }
 }
